[#579] Fix gcc conversion warnings in C++ runtime library

+++ compiler/extensions/cpp/runtime/src/zserio/BitBuffer.h
@@ -417,7 +417,7 @@ uint8_t BasicBitBuffer<ALLOC>::getMaskedLastByte() const
 
     return (lastByteBits == 0)
             ? m_buffer[roundedByteSize - 1]
-            : (m_buffer[roundedByteSize] & (0xFFU << (8U - lastByteBits)));
+            : static_cast<uint8_t>(m_buffer[roundedByteSize] & (0xFFU << (8U - lastByteBits)));
 }
 
 /** Typedef to BitBuffer provided for convenience - using std::allocator<uint8_t>. */
