+++ test/language/parameterized_types/cpp/PackedArrayElementParamTest.cpp
@@ -17,7 +17,7 @@ protected:
         for (uint16_t i = 0; i < numBlocks; ++i)
         {
             BlockHeader blockHeader;
-            blockHeader.setNumItems(i + 1);
+            blockHeader.setNumItems(static_cast<uint16_t>(i + 1));
             blockHeader.setOffset(0);
             headers.push_back(blockHeader);
         }
